% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcata.R
\encoding{UTF-8}
\name{adjust.brightness}
\alias{adjust.brightness}
\title{Adjust color brightness}
\usage{
adjust.brightness(rgb.in, percent = 10)
}
\arguments{
\item{rgb.in}{\code{rgb} of input color}

\item{percent}{the degree to which input color will be modified/brightened}
}
\value{
hex hex code for new color
}
\description{
Select suitable colors for highlighting plots.
}
\examples{
(rgb.in <- c(col2rgb("red")))
adjust.brightness(rgb.in, percent = 10)
}
