% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcata.R
\encoding{UTF-8}
\name{dist.city.block}
\alias{dist.city.block}
\title{Calculate city block distance between two matrices}
\usage{
dist.city.block(x, y)
}
\arguments{
\item{x}{first matrix}

\item{y}{second matrix}
}
\value{
cbdist city block distance between \code{x} and \code{y}
}
\description{
Calculates the city block distance between two matrices.
}
\examples{
  x <- matrix(0, nrow = 5, ncol = 7)
  y <- matrix(1, nrow = 5, ncol = 7)
  dist.city.block(x, y)

  y <- matrix(c(rep(0, 15), rep(1, 20)), nrow = 5, ncol = 7)
  dist.city.block(x, y)
}

