% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\encoding{UTF-8}
\name{lengthwhichis.na}
\alias{lengthwhichis.na}
\title{Count observations with missing data}
\usage{
lengthwhichis.na(x)
}
\arguments{
\item{x}{vector data which may contain missings}
}
\value{
\code{count} of observations where data are missing
}
\description{
Count observations with missing data.
}
\examples{
x <- c(rep(NA,18), rep(1,18), rep(0,10), rep(NA, 10))
lengthwhichis.na(x)
}

