% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcata.R
\encoding{UTF-8}
\name{make.palettes}
\alias{make.palettes}
\title{Convenience function for getting a pretty palette and highlight colours}
\usage{
make.palettes(n)
}
\arguments{
\item{n}{number of colours for each palette}
}
\value{
pal A character vector, \code{cv}, of colours that look pretty.

pal.light A character vector, \code{cv}, of matching highlight colours that look pretty.
}
\description{
Make a vector of n pretty colours, and n matching highlight colours.
}
\examples{
make.palettes(8)
}

