% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcata.R
\encoding{UTF-8}
\name{pretty_palette}
\alias{pretty_palette}
\title{Get a pretty palette of colours}
\usage{
pretty_palette(n)
}
\arguments{
\item{n}{number of colours in the palette}
}
\value{
cv A character vector, \code{cv}, of colours that look pretty.
}
\description{
Create a vector of n pretty colours.
}
\examples{
pretty_palette(8)
}
