% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempr.R
\docType{package}
\encoding{UTF-8}
\name{tempR}
\alias{tempR}
\title{tempR}
\description{
Analysis and visualization of data from temporal sensory methods, including for temporal check-all-that-apply (TCATA) and temporal dominance of sensations (TDS).
}
